Implement spectrum copy(), *, *= methods
In practice we will usually want __imul__, so makes sense to define
__mul__ using it. This requires a fresh Spectrum to write into. In
principle it would be more efficient to create one with an
"empty" data array...

... but as there is lots of code making modified array copies, life is
a lot simpler with a consistent copy() method used as a base for this
stuff. If copying the data becomes a noticeable bottleneck we could
add an `empty=False` argument that allows us to construct with
e.g. `y_data=np.empty_like(self.y_data)*ureg(self.y_data_unit)`, but
for now assume YAGNI!

With the new Python/numpy/Pint versions we can use np.copy() directly
on an array Quantity, which reduces unit-shuffling boilerplate.