Clean up lots of unnecessary casting between Quantity and magnitude
Newer versions of Pint and Numpy allow a lot more Numpy operations to
"just work" so we can cut a lot of boilerplate conversions.

Unfortunate np.linalg.norm is not one of these yet! For now we'll
leave the casting on that one, but np.square, np.sqrt and np.sum are
ok...