Remove _cell_volume private method
Now that Quantities are handled a bit better, it is really only
"needed" in force_constants... And it's not good practice to use
private methods from other modules anyway.

It doesn't hurt to make this casting to Bohr more explicit anyway, it
makes it clearer that this part of the maths is done in atomic units
without Pint wrappers.