Create a special action to upload version-limited Mac-ARM wheels
For 1.3.2 we built on x86 Mac (macos-13) but it would be good to
have (and test) some ARM wheels as well. We're having trouble building
those on older Python versions, but if we upload 3.10 and 3.11 wheels
to PyPI that should be somewhat useful.

This action is not intended to stick around once py3.8 and 3.9 are
deprecated; then we will try to build intel and ARM wheels in the same
job.