Release test against Numpy < 2
Troubleshooting test failures here... I *think* the wheels are all
built against pre-v2 numpy and so this test is failing due to the ABI
change when test runner uses a newer version.

In this case we can go ahead and make a new release that builds
against the newer Numpy ABI, which requires the min version to be
increased to 1.25.

One issue will be that Mantid is still on 1.24.