Bump numpy requirement to 1.24, matching Mantid (#309)
* Bump numpy requirement to 1.24, matching Mantid

While transitioning to Numpy 2.0 it is helpful to reduce the range of
versions being used and tested. As we are changing other requirements
in this release, it is a good time to bump Numpy.

Here we also update the build/release workflow to add Python 3.12 and
build with Numpy 2.0; these builds _should_ also work with all the
supported numpy 1.x.

* Bump threadpoolctl requirement to 3.0

"Min requirements" tests were throwing up some warnings suggesting that the openblas API wasn't being accessed succesfully. This seems to be resolved in version 3 and there are no knock-on effects to other dependencies.