Clean up warnings from Spectrum2D
To deal with broadening bin warnings, this uses an elegant approach to
error/warning "expectations" suggested by the pytest docs.

We find that different warning messages are raised by ostensibly the
same problem, depending on the code path. That's not ideal, but can
stay for now.

A unit-strip warning in test_mul is also fixed.