Clean up some low-hanging warnings in test suite (#310)
* Spglib 2.5 compatibility in brille.py

Spglib 2.5 introduces dataclasses for many of its data objects, and
raises a DeprecationWarning if you try to use it as a dictionary.

That's very nice, but we'd like to support a range of versions,
so for now we convert the new classes to a dict. We can migrate when
these versions of Spglib are more established.

* Suppress Spglib 2.5 deprecation warning raised by seekpath

Maybe when seekpath fixes this we can bump both version requirements?

* Refactor how warnings are checked in some tests

* Call to tight_layout() raises warning but seems to be unnecessary

* Explicitly fill masked array values with NaN

When this is implicit, a warning is raised.

* We find that different warning messages are raised by ostensibly the
same problem (irregular bin-width when broadening), depending on the code path. That's not ideal, but can
stay for now.

* Fix a unit-strip warning in test_mul functions