Move phonon-web-json exporter to separate file; implement --title
This is a fair chunk of code with a specific duty, and while it _can_
be used purely with QpointPhononModes as input is better characterised
as a function that uses both modes and a corresponding set of tick
labels. Move to a new "writers" module.

Use the existing CLI --title option to set title when using
euphonic-dispersion to generate JSON

Introduce a named type for XTickLabels; this helps legibility. There
is now some inconsistency between use of generic list/tuple and
classes from Typing: don't worry about it for now, this can be tidied
up afterwards.