Switch coverage checks to pytest-cov
pytest-cov uses Coverage but makes a few tweaks. The most important
change here is that it supports pytest-xdist by allowing each runner
to append information to .coverage. (This does make it more important
to delete .coverage occasionally...)

However, it also changes the workflow a bit for generating report
files. Some time was spent figuring out how to make that work!