Reduce source distribution size (#338)
## Update .gitattributes: exclude tests and workflows from archive

meson-python uses git-archive to create the sdist. Some of euphonic's
test files are on the large side and we don't need to include them
with every copy of the source.

This does have the downside that git-archive can no longer be used
to produce a copy of the source with tests, but we weren't using that
anyway.


## Build wheels from git checkout

While the purity of building from sdist was appealing, it is difficult
for cibuildwheel to cleanly run an external test suite as it
containerizes the Linux build (to deal with ManyLinux requirements).

We are still testing the sdist (on Windows) so just have to trust in
cibuildwheel's own efforts to separate the testing and source from the
installation. (Which it does seem to work quite hard at.)