Fix and test sdist builds; reduce sdist size (#337)
## Fix bug in build_utils/version.py

build_utils/version.py was raising errors in scenarios where git is unavailable. This happens more on Windows because 3 different names for git are attempted.

## Add windows sdist testing to pypi wheels workflow

Wheels are already tested by cibuildwheel but a separate action is now used to test the source build.

This test runs on Windows as this was the most vulnerable to the fixed bug

## Update .gitattributes: exclude tests, docs and workflows from archive

meson-python uses git-archive to create the sdist. Some of euphonic's
test files are on the large side and we don't need to include them
with every copy of the source.

This does have the downside that git-archive can no longer be used
to produce a copy of the source with tests, but we weren't using that
anyway.