Tweak TOX_SKIP_ENV logic: use dummy value
The trouble with ${{ test && '' || output }} is that '' is "false-y"
so it returns output even if the test fails.

The return value is treated as a regex for tox env to skip, so as
long as there are no matching envs we can use any string for the "run
all tests" case.