Update codecov configuration
We are hitting some rate-limiting, v4 with token should help with
that. It is also supposed to allow fork PRs to push coverage without a
token, although that is subject to rate-limiting.

The rate limit seems to be across all github projects using codecov,
which is why it fires off pretty frequently.