Avoid redundant coverage checks in CI (#353)
* Avoid pytest with coverage where unused

We only really need to use coverage for one pass over the test
suite, and only on the ubuntu runner (as this is the one that uploads
the coverage results.)

* Update codecov configuration

We are hitting some rate-limiting, v4 with token should help with
that. It is also supposed to allow fork PRs to push coverage without a
token, although that is subject to rate-limiting.

The rate limit seems to be across all github projects using codecov,
which is why it fires off pretty frequently.