Test 1D broadening consistency, pass restrict_range as needed
The current behaviour of Spectrum1D is to allow broadening with a
fixed-width on uneven bins (useful for rough q-broadening along a band
path) and to raise an error if variable-width broadening on those
bins.

This is inconsistent with Spectrum2D, which _does_ allow
variable-width broadening on irregular bins (and so allowed the
issue of inconsistent bin-edges to pass unnoticed...)

For now we preserve that restriction, but change the behaviour to
allow bin-edge extrapolation. As this broadening was previously
unavailable in such cases, it's not really an API break to get it
working again. (If your workflow relies on seeing this error message,
you might be the problem.)