Implement alternate _bin_centres_to_bin_edges mode (#363)
An optional parameter is provided to change how bin edges are
obtained from bin centres: previously the bin edges were constrained
to the initial data range, but this can lead to incorrect scaling
when performing broadening. Variable-width broadening schemes are
now allowed to extrapolate the bin edges in order to get the correct
width scaling.

Outside of broadening, the default behaviour is unchanged in order
to maintain backward compatibility.

The changes are tested and some other tweaks made to improve test coverage of these classes.