Bump CHANGELOG from set_version action
This could be done in a _slightly_ more modular fashion but would add
a bit of complexity as well to pass around the file changes across
actions without making multiple commits in the history.

The nice thing about this setup is that bump_changelog.py doesn't mind
at all if you already sorted out the CHANGELOG by hand, so if there
are any difficult/edge cases they can be handled by using a release
branch and running the script locally before using the release action.