Developing release action: split up build actions, no release yet
Change of plan: we _do_ need to push the version/tag commit in order
to have a sensible multi-job workflow. We do want that multi-job
workflow because versioning tweaks should happen on one OS but
building/testing needs to be matrixed.

But this should generally be done on a release branch, so if the
release fails it will be possible to clean up these commits/tags
without messing up the master branch history.

It seems a good idea to provide standalone build/test actions for
debugging. By separating the sdist and wheel builds they can be called
in parallel as part of overall release workflow.

A couple of details still to work out at this point:

- The OS matrix should be moved up to release.yml and passed as an
  option to build_wheels, so it is possible to troubleshoot one OS at a
  time.

- We may need to store the normalised release tag to environmnent and use
  in subsequent steps instead of inputs.ref. That way, small formatting
  changes like adding "v" will be carried forward correctly.

  https://docs.github.com/en/actions/sharing-automations/reusing-workflows#using-outputs-from-a-reusable-workflow