Maybe swapping these steps will improve Github visualisation
The workflow visualisation tends to put the "test sdist" step right on
top of the dependency line from build-wheels to release.

This is quite misleading as it looks like test sdist depends on
build-wheels and yet it can start running before build-wheels is done.

Perhaps if we swap the jobs they will be stacked in a different order
and it will look clearer?