Fix incompatible-pointer-types (#378)
* Correct tox version for CI extra

* Pointer casting to PyObject for GCC14 compliance

Euphonic was failing to build with GCC-14.

Compilers are understandably twitchy about the ensuing signature mismatch when we take a function with signature

  Pyobject *PyDict_GetItemString(PyObject *p, const char *key)

and pass it a *PyDictObject as the first argument. Since GCC14 this has become a hard error; the solution is to explicitly cast pointers from their specialised type to *PyObject.

* Update CHANGELOG

* Add Github actions workflow for Ubuntu 24.04 / GCC-14