Linting: whitespace, negation, type comparison, mutable defaults, staticmethods (#379)
* Linting: prefer isinstance to exact type check (C0123)

* Linting: remove trailing whitespace (C0303)

* Linting: unnecessary negation (C0117)

* Use None instead of dangerous mutable default arguments

It is very bad practice to use a mutable default argument like [] or
{}. The object are defined when the function is created, and shared
among function calls; this allows state to unexpectedly leak between
functions if they are mutated.

* Methods without arguments should be @staticmethod (E0211)

Actually these ones should probably by pytest fixtures but let's leave
that for another day.

* Fix type-checking logic

QpointPhononModes is a subclass of QpointFrequencies
The intent of these checks was to see if the class is
QpointFrequencies _and not_ QpointPhononModes. Let's just check if it
is a FC or QPM instead.

* Update euphonic/io.py

Co-authored-by: Jacob Wilkins <46597752+oerc0122@users.noreply.github.com>

* Apply suggestions from code review

Co-authored-by: Jacob Wilkins <46597752+oerc0122@users.noreply.github.com>

* Fix missing import

* Apply suggestions from code review

Co-authored-by: Jacob Wilkins <46597752+oerc0122@users.noreply.github.com>

* More whitespace / mutable default fixes found by ruff

* Minimal ruff configuration

This can be expanded as more linting is completed

* Add a linting prerequisite step to tests

* Call GCC14 test from run-tests: this gates it behind lint step

---------

Co-authored-by: Jacob Wilkins <46597752+oerc0122@users.noreply.github.com>