Exempt tests from RUF012
This rule is supposed to encourage you to use immutable types for
class variables, and put mutable stuff in __init__... or at least use
a ClassVar annotation to indicate that instance methods should stay
away.

But in pytests it is often useful to create a big block of dicts and
lists for test parametrisation, and it is clearest to do it just above
the test definition. So this practice just gets in the way, for little
benefit.