Remove (generally) unreachable error from CASTEP phonon_dos reader
As far as I can tell CASTEP will never produce a .phonon_dos file with
gradients but no DOS. A regular .phonon file will hit a TypeError
before it reaches this line.

We can only test the error with a specially crafted corrupt
.phonon_dos file. There are infinite other possible data corruptions
that don't have dedicated error messages.

So it is not useful and should be removed.