CI troubleshooting (#404)
min-requirements tests seem to have broken overnight. We never did get to the bottom of why it is broken _now_, but the fix is something we've been meaning to do for a while, anyway:

* Bump spglib to 2.1.0

This version included an overhaul of build/packaging, which seems a safer target for our minimum support.