Linting: flake8-errmsg (#402)
* Apply flake8-errmsg

Use a dedent_and_fill function to streamline error-message writing: nice indented triple-quoted blocks in the source can be automatically realigned and line-wrapped for display to user.

* Drop unnecessary encoding line from some files

* Give module-scope err msg a different name from function-level

Be clear that we are not trying to reuse or clobber this variable

* Refactor broadening test for reusable data

* Improve test coverage of error messages

* Remove some unreachable/untestable error messages

* Use mocking to produce spglib error reliably

* Test 2D Lorentz broadening on one axis only

This is a different logic path to the existing Gaussian (and both-axes
Lorentzian) cases.

---------

Co-authored-by: Jacob Wilkins <46597752+oerc0122@users.noreply.github.com>