Add unit test for dipole-dipole subtraction with scrambled sc_matrix
This MgO case has a non-symmetric supercell matrix. Rather than an
integration test over the whole Phonopy import process we can localise
the regression test to this force_constants data.

Note that a ForceConstants JSON is being used as a convenient
collection of input data to from_total_fc_with_dipole: it is _not_ a
generally valid Euphonic force constants JSON file as the force
constants are inconsistent with the dipole data.