Drop redundant transpose operations
As

  (M^-1).T = (M.T)^-1

the second transpose simply reverses the first one. Our choice should
then be based on consistency, convention and legibility

It doesn't seem right that "sc_to_u_matrix" would involve
transposition relative to sc_matrix, it should be a directly analogous.
The definition of SC involves transposition but that is related to the
conventional orientation of the lattice vectors:
https://phonopy.github.io/phonopy/phonopy-module.html#supercell-matrix

These conventions are a bit of a headache. I don't think this
transpose-free option is _less_ justified than the existing
arrangement, and it involves less code.