Numpy tidying up: replace 2-D einsums with @, newAxis with None
Not that there's much wrong with einsum, but this makes the dot
products a bit easier to scan past as "nothing complicated happening
here" and concentrate on the more exotic ones. There can also be a
slightl performance boost as matmul will dispatch to optimal BLAS
routines more reliably than einsum.

Similarly, there is nothing wrong with using np.newAxis but here there
are a few places it was aliased to "ax" to save space and that is a
lot more confusing than using None!