Bugfix: calculation of atom positions during dipole-dipole subtraction (#417)
* Bugfix

Issue #377 Reported some inconsistencies between Phonopy and Euphonic band
structures when using long-range corrections. The system at issue has
a rotation between unit cell and primitive cell, which leads
to a non-symmetric derived supercell matrix. That makes it rather easy
for a missing/extra transpose() operation to cause subtle problems!

* Unit test

A MgO unit test case is created with a non-symmetric supercell matrix. Rather than an
integration test over the whole Phonopy import process, we can localise
the regression test to this force_constants data.

Note that a ForceConstants JSON is being used as a convenient
collection of input data to from_total_fc_with_dipole: it is _not_ a
generally valid Euphonic force constants JSON file as the force
constants are inconsistent with the dipole data.

* Replace some einsum with cleaner (faster?) matmul

These einsum are just doing dot products, there is nicer syntax for
that. It should mean that the operation also gets offloaded to
optimised BLAS routines more reliably. (But this step is unlikely to
be much of a bottleneck, so the main motivation is readability.)