Detect OpenMP library conflicts and fallback to serial calculation (#432)
Detect when multiple OpenMP libraries are active, and force serial Euphonic in this case. This defends against some nasty crashes and hangs, especially seen on MacOS where Euphonic wheels use homebrew OpenMP and Brille wheels package their own OpenMP from Conan.

A warning is raised when these conflicts are seen, and this is unit tested with a bit of mocking.

Brille imports are deferred in the test suite so that pytest does not needlessly
import Brille if we are not going to run any Brille tests. (i.e. when using `run_tests.py -m 'not brille'`)
This seems to result in a test suite that emits warnings if you run everything at once (and load multiple OpenMP in the process), but not if you run separate `-m brille` and `-m 'not brille'` separately.

Note that some segfaults were encountered when testing Euphonic with Brille, even with this OpenMP logic in place. They seem to be unrelated, and fixed by https://github.com/brille/brille/pull/113

* Update CHANGELOG