Overhaul handling of RNG: pass generators explicitly
This requires less monkey-patching at test time and gives the user
more opportunity to use fixed seeds or quasirandom generators

A monkey-patch is still used to handle the show_sampling CLI tool
tests: this tool doesn't provide user input for seeds so a
well-defined hook for patching seems a reasonably clean solution.

Surprisingly, it turns out that test_powder_map doesn't actually need
a fixed random seed because none of the tests use jitter. That's
possibly a bad thing but for now let's roll with it and use that to
simplify further. If we introduce some RNG-sensitive tests we can use
the same "clear monkeypatch hook" strategy as test_sampling.